feat(back): add business rule for subtracting stock amount on new order save
Had to change stockAmount type from int to Long to match product
quantity

+++ setec-backend/crud/src/main/java/com/setec/crud/dto/ProductRequest.java
@@ -5,6 +5,6 @@ import jakarta.validation.constraints.NotBlank;
 public record ProductRequest(
         @NotBlank String description,
         float price,
-        int stockAmount
+        Long stockAmount
 ) {
 }
